move functions and tests into package directory
